requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-session-timeout==0.1.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.1
psutil==5.9.1
python-decouple==3.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
cryptography

